requirements.txt
dj-database-url==0.5.0
gunicorn==19.6.0
Django==2.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-autodocs-0.4.4
